chore: rename ToolTip component
